src/prj/requirements.txt
asgiref==3.8.1
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
diff-match-patch==20241021
Django==5.1.4
django-filter==24.3
django-import-export==4.3.4
docker==7.1.0
et_xmlfile==2.0.0
folium==0.19.4
fonttools==4.55.3
geopandas==1.0.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mysqlclient==2.2.7
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.3.0
xyzservices==2024.9.0
openpyxl==3.1.5